Refactor + docs
